Record timestamp on detection and timed expiry
With short detection windows devices are missed.

If you have a short detection window and longer expiry time,
it is more responsive to join events.  With a shorter expiry
time, leave events are more responsive but less reliable.

By default, update every 15 seconds and mark expired if not
detected in 60 seconds.